+++ oneIdentity/src/app/handlers/oidcAuthorize/utils/AwsParameters.ts
