+++ oneIdentity/src/app/handlers/oidcAuthorize/utils/Object.ts
