+++ oneIdentity/src/app/handlers/oidcAuthorize/utils/PublicKey.ts
@@ -1,5 +1,5 @@
 import jwkToPem from "jwk-to-pem";
-import { JWKS } from "../../models/Jwks";
+import { JWKS } from "../models/Jwks";
 import { ValidationException } from "../exception/validationException";
 import { get, isCacheActive } from "./Jwks/JwksCache";
 import { getJwks } from "./Jwks/JwksRetriever";
