+++ oneIdentity/src/app/handlers/oidcAuthorize/utils/String.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyEventHeaders } from "aws-lambda";
-import { SPID_FISCAL_NUMBER_PREFIX } from "../../models/Token";
+import { SPID_FISCAL_NUMBER_PREFIX } from "../models/Token";
 
 /**
  * Converts all header keys to lowercase.
