+++ oneIdentity/src/test/utils/Object.test.ts
@@ -1,4 +1,4 @@
-import { copyAndMaskObject } from "../../app/utils/Object";
+import { copyAndMaskObject } from "../../app/handlers/oidcAuthorize/utils/Object";
 import { oneIdentityIdTokenMock } from "../__mock__/token.mock";
 
 describe("Object utils Tests", () => {
