+++ oneIdentity/src/test/validation/Origin.test.ts
@@ -1,4 +1,4 @@
-import { isOriginAllowed } from "../../app/validation/Origin";
+import { isOriginAllowed } from "../../app/handlers/oidcAuthorize/validation/Origin";
 import { setupEnv } from "../test.utils";
 
 describe("Origin Validation tests", () => {
